requirements.txt
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
dj-database-url==0.5.0
Django==3.1.1
django-cities==0.6
django-cors-headers==3.5.0
django-countries==6.1.3
django-heroku==0.3.1
django-js-asset==1.2.2
django-localflavor==3.0.1
django-model-utils==4.0.0
django-taggit==1.3.0
djangorestframework==3.12.1
flake8==3.8.3
gunicorn==20.0.4
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.6.0
python-stdnum==1.14
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
swapper==1.1.2.post1
toml==0.10.1
tqdm==4.50.0
whitenoise==5.2.0
wrapt==1.12.1
